GH-45254: [C++][Acero] Fix the row offset truncation in row table merge (#45255)
### Rationale for this change

See #45254

### What changes are included in this PR?

First modify the test case to expose the suspecting bug.

Then the fix in source.

### Are these changes tested?

By existing tests.

### Are there any user-facing changes?

None.

* GitHub Issue: #45254

Authored-by: Rossi Sun <zanmato1984@gmail.com>

+++ cpp/src/arrow/acero/swiss_join.cc
@@ -439,11 +439,11 @@ Status RowArrayMerge::PrepareForMerge(RowArray* target,
     num_rows = 0;
     num_bytes = 0;
     for (size_t i = 0; i < sources.size(); ++i) {
-      target->rows_.mutable_offsets()[num_rows] = static_cast<uint32_t>(num_bytes);
+      target->rows_.mutable_offsets()[num_rows] = num_bytes;
       num_rows += sources[i]->rows_.length();
       num_bytes += sources[i]->rows_.offsets()[sources[i]->rows_.length()];
     }
-    target->rows_.mutable_offsets()[num_rows] = static_cast<uint32_t>(num_bytes);
+    target->rows_.mutable_offsets()[num_rows] = num_bytes;
   }
 
   return Status::OK();
